fix: mobile layout, draggable vertices, lock view UX
Mobile layout (all 3 map tabs)
- Single scrollable column on phones; the parent `<main>` handles the scroll
  instead of nested containers fighting for space.
- Map gets a fixed `h-[50vh]` on mobile, `lg:h-full` on desktop.
- Sidebar overflow is gated behind `lg:` so its content flows in the parent
  scroll on mobile instead of being clipped.
- Identity tab adds a mobile-only address combobox pinned ABOVE the map
  (last in DOM + `flex-col-reverse` to surface it visually).

Perimeter vertices
- Each vertex is now a draggable AdvancedMarker (live drag + drop-end
  refresh of the polygon area).
- Shift+click removes a vertex.
- `draggingRef` flag swallows the trailing map "click" that Google fires on
  drag mouseup — no more phantom vertex at the drop point.
- Hover cursor: pointer; active drag: grabbing.

Poles
- Same `draggingRef` flag prevents a phantom pole appearing where the user
  releases an existing one.

Identity / lock-view
- `lockedZoom` schema relaxed from int to float — stored as the user's
  exact camera zoom so the reloaded framing matches their original view
  (was clipping the polygon when reloading after a non-integer zoom).
- "Modifier la vue" resets perimeter + poles unconditionally (their
  positions are anchored to the prior framing; the confirm dialog was
  removed for a more direct flow).

Cursor
- Crosshair forced only on `.gm-style` root (not on every descendant) so
  vertex / pole markers can declare their own pointer/grabbing cursor.

+++ lib/estimation/schema.ts
@@ -169,8 +169,12 @@ export const fieldSchema = z.object({
    * Map zoom level the user "froze" the view at. Once set, the perimeter and
    * poles tabs render a non-interactive Maps JS view at this zoom + the
    * field's lat/lng. Unset until the user clicks "Confirmer la vue".
+   *
+   * Stored as a float (Google Maps allows fractional zoom). Previously we
+   * `Math.round`-ed this value, which caused the reloaded framing to differ
+   * from what the user originally saw, sometimes clipping the polygon.
    */
-  lockedZoom: z.number().int().min(1).max(21).optional(),
+  lockedZoom: z.number().min(1).max(21).optional(),
   // Perimeter — polygon vertices in real-world lat/lng
   perimeter: z.array(geoPointSchema).optional(),
   surfaceM2: z.coerce.number().nonnegative("Must be ≥ 0").max(500000),
